refactor: move avatar to another component
