+++ src/components/Comments/Avatar.tsx
@@ -1,11 +1,31 @@
 import React from 'react'
+import styled from 'styled-components'
+
+const Wrapper = styled.div`
+  position: absolute;
+  margin-left: -60px;
+  img {
+    border-radius: 40px;
+  }
+`
 
 interface AvatarProps {
-  commentsUrl: string
+  readonly url: string
+  readonly avatar: string
+  readonly login: string
 }
 
-const Avatar: React.FC<AvatarProps> = () => {
-  return <h1>reactions</h1>
-}
-
+const Avatar: React.FC<AvatarProps> = ({url, avatar, login}) => (
+  <Wrapper>
+    <a href={url} title={`@${login}`}>
+      <img
+        className="avatar avatar-user"
+        height="40"
+        width="40"
+        alt={`@${login}`}
+        src={avatar}
+      />
+    </a>
+  </Wrapper>
+)
 export default Avatar
