+++ src/components/Comments/Comments.tsx
@@ -1,6 +1,7 @@
 import React from 'react'
 import styled from 'styled-components'
 import {getTimeAgoInWords} from '../../utils'
+import Avatar from './Avatar'
 
 const Comment = styled('div')`
   margin-top: 16px;
@@ -13,11 +14,6 @@ const Comment = styled('div')`
     background: none;
   }
 
-  .avatar {
-    position: absolute;
-    margin-left: -60px;
-    border-radius: 40px;
-  }
   .content {
     border: 1px solid #efefef;
     border-radius: 6px;
@@ -80,24 +76,6 @@ const Actions = styled('div')`
   //margin-bottom: 40px;
   float: right;
 `
-
-interface AvatarProps {
-  readonly url: string
-  readonly avatar: string
-  readonly login: string
-}
-
-const Avatar: React.FC<AvatarProps> = ({url, avatar, login}) => (
-  <a href={url} title={`@${login}`}>
-    <img
-      className="avatar avatar-user"
-      height="40"
-      width="40"
-      alt={`@${login}`}
-      src={avatar}
-    />
-  </a>
-)
 
 type ReactionKey =
   | '+1'
@@ -178,91 +156,93 @@ const Comments: React.FC<CommentsProps> = ({url, comments}) => {
   return (
     <>
       <h2>Comments</h2>
+      <p>Thanks for reading ❤️ </p>
+
+      <p>
+        {' '}
+        You can comment by replying to the{' '}
+        <a href={url}>issue for this post.</a>
+        {` `}
+      </p>
+
+      {(!comments || comments.length === 0) && (
         <p>
-          Thanks for reading ❤️        </p>
-
-          <p> You can comment by replying to the{' '}
-          <a href={url}>issue for this post.</a>{` `}
-          </p>
-
-          {(!comments || comments.length === 0) && (
-            <p>
-              There’s no comments yet,{` `}
-              <a
-                href={`${url}#new_comment_field`}
-                target="_blank"
-                rel="noreferrer noopener"
-                className="button btn"
-              >
-                be the first to leave one!
-              </a>
-            </p>
-          )}
-        {comments &&
-          // eslint-disable-next-line @typescript-eslint/no-explicit-any
-          comments.map((comment: any) => {
-            return (
-              <Comment key={comment.id} className="comment sans">
-                <div className="content">
-                  <Avatar
-                    url={comment.user.html_url}
-                    avatar={comment.user.avatar_url}
-                    login={comment.user.login}
-                  />
-                  <div className="header">
-                    <div className="date" data-date={comment.created_at}>
-                      <a href={comment.user.html_url}>
-                        <strong>@{comment.user.login}</strong>
-                      </a>
-                      {` `}
-                      commented{' '}
-                      <a href={comment.html_url}>
-                        {getTimeAgoInWords(comment.created_at)} ago
-                      </a>
-                      {` `}
-                      {comment.updated_at > comment.created_at && (
-                        <span>• edited</span>
-                      )}
-                    </div>
-                    {comment.author_association === 'OWNER' && (
-                      <div className="association">Author</div>
-                    )}
-                    {comment.author_association === 'OWNER' && (
-                      <div className="association">Owner</div>
+          There’s no comments yet,{` `}
+          <a
+            href={`${url}#new_comment_field`}
+            target="_blank"
+            rel="noreferrer noopener"
+            className="button btn"
+          >
+            be the first to leave one!
+          </a>
+        </p>
+      )}
+      {comments &&
+        // eslint-disable-next-line @typescript-eslint/no-explicit-any
+        comments.map((comment: any) => {
+          return (
+            <Comment key={comment.id} className="comment sans">
+              <div className="content">
+                <Avatar
+                  url={comment.user.html_url}
+                  avatar={comment.user.avatar_url}
+                  login={comment.user.login}
+                />
+                <div className="header">
+                  <div className="date" data-date={comment.created_at}>
+                    <a href={comment.user.html_url}>
+                      <strong>@{comment.user.login}</strong>
+                    </a>
+                    {` `}
+                    commented{' '}
+                    <a href={comment.html_url}>
+                      {getTimeAgoInWords(comment.created_at)} ago
+                    </a>
+                    {` `}
+                    {comment.updated_at > comment.created_at && (
+                      <span>• edited</span>
                     )}
                   </div>
-                  <div
-                    className="markdown-body sans"
-                    dangerouslySetInnerHTML={{__html: comment.body_html}}
-                  />
-                  <div className="reactions">
-                    {REACTIONS.map(reaction => (
-                      <Reaction
-                        key={reaction}
-                        url={comment.html_url}
-                        reaction={reaction as ReactionKey}
-                        count={+comment.reactions[reaction]}
-                      />
-                    ))}
-                    <a
-                      href={comment.html_url}
-                      title="Add a reaction"
-                      target="_blank"
-                      rel="noreferrer noopener"
-                      className="react"
-                    >
-                      <div>
-                        <svg
-                          className="octicon octicon-smiley"
-                          viewBox="0 0 16 16"
-                          version="1.1"
-                          width="16"
-                          height="16"
-                          aria-hidden="true"
-                        >
-                          <path
-                            fillRule="evenodd"
-                            d="M1.5 8a6.5 6.5 0 1113 0 6.5 6.5 0 01-13 0zM8 0a8 8 0 100 16A8 8 0 008 0zM5 8a1 1 0 100-2 1 1 0 000
+                  {comment.author_association === 'OWNER' && (
+                    <div className="association">Author</div>
+                  )}
+                  {comment.author_association === 'OWNER' && (
+                    <div className="association">Owner</div>
+                  )}
+                </div>
+                <div
+                  className="markdown-body sans"
+                  dangerouslySetInnerHTML={{__html: comment.body_html}}
+                />
+                <div className="reactions">
+                  {REACTIONS.map(reaction => (
+                    <Reaction
+                      key={reaction}
+                      url={comment.html_url}
+                      reaction={reaction as ReactionKey}
+                      count={+comment.reactions[reaction]}
+                    />
+                  ))}
+                  <a
+                    href={comment.html_url}
+                    title="Add a reaction"
+                    target="_blank"
+                    rel="noreferrer noopener"
+                    className="react"
+                  >
+                    <div>
+                      <svg
+                        className="octicon octicon-smiley"
+                        viewBox="0 0 16 16"
+                        version="1.1"
+                        width="16"
+                        height="16"
+                        aria-hidden="true"
+                      >
+                        <path
+                          fillRule="evenodd"
+                          d="M1.5 8a6.5 6.5 0 1113 0 6.5 6.5 0 01-13 0zM8 0a8 8 0 100 16A8 8 0 008 0zM5 8a1 1 0 100-2 1 1 0 000
                             2zm7-1a1 1 0 11-2 0 1 1 0 012 0zM5.32 9.636a.75.75 0
                             011.038.175l.007.009c.103.118.22.222.35.31.264.178.683.37 1.285.37.602 0 1.02-.192
                             1.285-.371.13-.088.247-.192.35-.31l.007-.008a.75.75 0
@@ -270,25 +250,25 @@ const Comments: React.FC<CommentsProps> = ({url, comments}) => {
                             1.984 0 01-.184.213c-.16.166-.338.316-.53.445-.63.418-1.37.638-2.127.629-.946
                             0-1.652-.308-2.126-.63a3.32 3.32 0
                             01-.715-.657l-.014-.02-.005-.006-.002-.003v-.002h-.001l.613-.432-.614.43a.75.75 0 01.183-1.044h.001z"
-                          />
-                        </svg>
-                      </div>
-                    </a>
-                  </div>
+                        />
+                      </svg>
+                    </div>
+                  </a>
                 </div>
-              </Comment>
-            )
-          })}
-        <Actions>
-          <a
-            href={`${url}#new_comment_field`}
-            target="_blank"
-            rel="noreferrer noopener"
-            className="button btn"
-          >
-            Add a comment &rarr;
-          </a>
-        </Actions>
+              </div>
+            </Comment>
+          )
+        })}
+      <Actions>
+        <a
+          href={`${url}#new_comment_field`}
+          target="_blank"
+          rel="noreferrer noopener"
+          className="button btn"
+        >
+          Add a comment &rarr;
+        </a>
+      </Actions>
     </>
   )
 }
